# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crm",
        "description": "crm vue jwt",
        "contact": {
            "email": "info@digex"
        },
        "license": {
            "name": "crm",
            "url": "http://crm.loc/"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "crm  host server"
        }
    ],
    "paths": {
...
        "/api/customers/index": {
            "get": {
                "tags": [
                    "customers"
                ],
                "description": "get custumers",
                "operationId": "App\\Http\\Controllers\\CustomerController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "get by type [COMPANIES,PERSONS]",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/customers/search": {
            "get": {
                "tags": [
                    "customers"
                ],
                "operationId": "App\\Http\\Controllers\\CustomerController::search",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "name  of  customer persons",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/incomes/categories/index": {
            "get": {
                "tags": [
                    "incomes"
                ],
                "description": "get index",
                "operationId": "App\\Http\\Controllers\\InvoiceCategoryController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/incomes/categories/get": {
            "get": {
                "tags": [
                    "incomes"
                ],
                "description": "get company by id",
                "operationId": "App\\Http\\Controllers\\InvoiceCategoryController::get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id of category",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/incomes/categories/sync": {
            "post": {
                "tags": [
                    "incomes"
                ],
                "description": "get company by id",
                "operationId": "App\\Http\\Controllers\\InvoiceCategoryController::sync",
                "parameters": [
                    {
                        "name": "item",
                        "in": "query",
                        "description": "item",
                        "required": false,
                        "schema": {
                            "properties": {
                                "item[name]": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/incomes/categories/destroy/{item}": {
            "delete": {
                "tags": [
                    "incomes"
                ],
                "description": "get company by id",
                "operationId": "App\\Http\\Controllers\\InvoiceCategoryController::destroy",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "id of category",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
...
        "/api/items/index": {
            "get": {
                "tags": [
                    "items"
                ],
                "description": "get companies",
                "operationId": "App\\Http\\Controllers\\ItemController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/items/get": {
            "get": {
                "tags": [
                    "items"
                ],
                "description": "get companies",
                "operationId": "App\\Http\\Controllers\\ItemController::get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id of category",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "auth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}